code/requirements.txt

ipython
click
scandir
boto3
ujson
matplotlib
jupyter
pandas
us
iso8601
attrs
